+++ __tests__/location.ts
@@ -17,6 +17,7 @@ test("decodeLocation parses well-formed paths properly", () => {
   const equal = getEqual(false);
 
   equal("/", "/", {
+    key: "1vr879e-0",
     path: [],
     search: {},
     raw: {
@@ -27,6 +28,7 @@ test("decodeLocation parses well-formed paths properly", () => {
   });
 
   equal("/test", "/test", {
+    key: "15r3z8l-0",
     path: ["test"],
     search: {},
     raw: {
@@ -37,6 +39,7 @@ test("decodeLocation parses well-formed paths properly", () => {
   });
 
   equal("/test/ID", "/test/ID", {
+    key: "1t85fsm-0",
     path: ["test", "ID"],
     search: {},
     raw: {
@@ -47,6 +50,7 @@ test("decodeLocation parses well-formed paths properly", () => {
   });
 
   equal("/test/repositories/mine", "/test/repositories/mine", {
+    key: "2tnzaa-0",
     path: ["test", "repositories", "mine"],
     search: {},
     raw: {
@@ -57,6 +61,7 @@ test("decodeLocation parses well-formed paths properly", () => {
   });
 
   equal("/test#foo", "/test#foo", {
+    key: "15r3z8l-vqe7l8",
     path: ["test"],
     search: {},
     hash: "foo",
@@ -71,6 +76,7 @@ test("decodeLocation parses well-formed paths properly", () => {
     "/profile/settings?invitation=code",
     "/profile/settings?invitation=code",
     {
+      key: "hvvndr-1mfzit5",
       path: ["profile", "settings"],
       search: { invitation: "code" },
       raw: {
@@ -82,6 +88,7 @@ test("decodeLocation parses well-formed paths properly", () => {
   );
 
   equal("/test?filters=foo&filters=bar", "/test?filters=foo&filters=bar", {
+    key: "15r3z8l-1iu5pw9",
     path: ["test"],
     search: { filters: ["foo", "bar"] },
     raw: {
@@ -96,6 +103,7 @@ test("decodeLocation parses route with empty params properly", () => {
   const equal = getEqual(false);
 
   equal("/?#", "/#", {
+    key: "1vr879e-wm3gqk",
     path: [],
     search: {},
     hash: "",
@@ -107,6 +115,7 @@ test("decodeLocation parses route with empty params properly", () => {
   });
 
   equal("/test?", "/test", {
+    key: "15r3z8l-0",
     path: ["test"],
     search: {},
     raw: {
@@ -117,6 +126,7 @@ test("decodeLocation parses route with empty params properly", () => {
   });
 
   equal("/test?foo", "/test?foo", {
+    key: "15r3z8l-1xxto3f",
     path: ["test"],
     search: { foo: "" },
     raw: {
@@ -127,6 +137,7 @@ test("decodeLocation parses route with empty params properly", () => {
   });
 
   equal("/test?foo=", "/test?foo", {
+    key: "15r3z8l-1xxto3f",
     path: ["test"],
     search: { foo: "" },
     raw: {
@@ -137,6 +148,7 @@ test("decodeLocation parses route with empty params properly", () => {
   });
 
   equal("/test?foo=&foo", "/test?foo&foo", {
+    key: "15r3z8l-d9a9pr",
     path: ["test"],
     search: { foo: ["", ""] },
     raw: {
@@ -147,6 +159,7 @@ test("decodeLocation parses route with empty params properly", () => {
   });
 
   equal("/test#", "/test#", {
+    key: "15r3z8l-wm3gqk",
     path: ["test"],
     search: {},
     hash: "",
@@ -162,6 +175,7 @@ test("decodeLocation parses URI components (and encode it back for url)", () => 
   const equal = getEqual(false);
 
   equal("/:test+", "/%3Atest%2B", {
+    key: "uo6seg-0",
     path: [":test+"],
     search: {},
     raw: {
@@ -172,6 +186,7 @@ test("decodeLocation parses URI components (and encode it back for url)", () => 
   });
 
   equal("/test??foo", "/test?%3Ffoo", {
+    key: "15r3z8l-md9mo5",
     path: ["test"],
     search: { "?foo": "" },
     raw: {
@@ -182,6 +197,7 @@ test("decodeLocation parses URI components (and encode it back for url)", () => 
   });
 
   equal("/test?foo=?", "/test?foo=%3F", {
+    key: "15r3z8l-5fvrh2",
     path: ["test"],
     search: { foo: "?" },
     raw: {
@@ -192,6 +208,7 @@ test("decodeLocation parses URI components (and encode it back for url)", () => 
   });
 
   equal("/test##+foo", "/test#%23%2Bfoo", {
+    key: "15r3z8l-187n8mo",
     path: ["test"],
     search: {},
     hash: "#+foo",
@@ -207,6 +224,7 @@ test("decodeLocation keeps extra slashes when removeExtraSlashes is false", () =
   const equal = getEqual(false);
 
   equal("///", "///", {
+    key: "1chjp8o-0",
     path: ["", "", ""],
     search: {},
     raw: {
@@ -217,6 +235,7 @@ test("decodeLocation keeps extra slashes when removeExtraSlashes is false", () =
   });
 
   equal("/test/", "/test/", {
+    key: "1xcznd3-0",
     path: ["test", ""],
     search: {},
     raw: {
@@ -227,6 +246,7 @@ test("decodeLocation keeps extra slashes when removeExtraSlashes is false", () =
   });
 
   equal("/test/?foo", "/test/?foo", {
+    key: "1xcznd3-1xxto3f",
     path: ["test", ""],
     search: { foo: "" },
     raw: {
@@ -240,6 +260,7 @@ test("decodeLocation keeps extra slashes when removeExtraSlashes is false", () =
     "//profile//settings?invitation=/",
     "//profile//settings?invitation=%2F",
     {
+      key: "1nmu3h0-2e1b3u",
       path: ["", "profile", "", "settings"],
       search: { invitation: "/" },
       raw: {
@@ -255,6 +276,7 @@ test("decodeLocation removes extra slashes when removeExtraSlashes is true", () 
   const equal = getEqual(true);
 
   equal("///", "/", {
+    key: "1vr879e-0",
     path: [],
     search: {},
     raw: {
@@ -265,6 +287,7 @@ test("decodeLocation removes extra slashes when removeExtraSlashes is true", () 
   });
 
   equal("/test/", "/test", {
+    key: "15r3z8l-0",
     path: ["test"],
     search: {},
     raw: {
@@ -275,6 +298,7 @@ test("decodeLocation removes extra slashes when removeExtraSlashes is true", () 
   });
 
   equal("/test/?foo", "/test?foo", {
+    key: "15r3z8l-1xxto3f",
     path: ["test"],
     search: { foo: "" },
     raw: {
@@ -288,6 +312,7 @@ test("decodeLocation removes extra slashes when removeExtraSlashes is true", () 
     "//profile//settings?invitation=/",
     "/profile/settings?invitation=%2F",
     {
+      key: "hvvndr-2e1b3u",
       path: ["profile", "settings"],
       search: { invitation: "/" },
       raw: {
