+++ package.json
@@ -46,6 +46,7 @@
     "react": ">=16.8.0"
   },
   "dependencies": {
+    "@emotion/hash": "^0.8.0",
     "history": "^5.3.0",
     "use-sync-external-store": "1.0.0-rc.2 || ^1.0.0"
   },
