+++ src/location.ts
@@ -1,3 +1,4 @@
+import mumurhash from "@emotion/hash";
 import { Path as HistoryLocation } from "history";
 import { isNonEmpty } from "./helpers";
 import { decodeSearch, encodeSearch } from "./search";
@@ -27,6 +28,8 @@ export const decodeLocation = (
   const stringifiedLocation = rawPath + rawSearch + rawHash;
 
   return {
+    key: `${mumurhash(rawPath)}-${mumurhash(rawSearch + rawHash)}`,
+
     path: parsedPath,
     search: parsedSearch,
     ...(parsedHash !== null && {
