+++ src/matcher.ts
@@ -139,31 +139,15 @@ export const extractLocationParams = (
   return params;
 };
 
-export const matchAll = (
-  location: Location,
-  matchers: Matcher[],
-): { name: string; params: Params }[] => {
-  const routes = [];
-  for (const matcher of matchers) {
-    const params = extractLocationParams(location, matcher);
-
-    if (params != null) {
-      routes.push({ name: matcher.name, params });
-    }
-  }
-
-  return routes;
-};
-
 export const match = (
   location: Location,
   matchers: Matcher[],
-): { name: string; params: Params } | undefined => {
+): { key: string; name: string; params: Params } | undefined => {
   for (const matcher of matchers) {
     const params = extractLocationParams(location, matcher);
 
     if (params != null) {
-      return { name: matcher.name, params };
+      return { key: location.key, name: matcher.name, params };
     }
   }
 };
